+++ CHANGELOG.md
@@ -6,7 +6,12 @@ unreleased
 - `dropinInstance.requestPaymentMethod` will return a promise if no callback is provided
 - `dropinInstance.teardown` will return a promise if no callback is provided
 - `dropin.create` will return a promise if no callback is provided
-- Use npm scoped version of browser-detection
+
+1.3.1
+-----
+- Use version 3.19.0 of braintree-web
+- Autoprefix CSS, fixing issues in older browsers
+- Add aria-invalid attribute for cards
 
 1.3.0
 ------
@@ -16,6 +21,7 @@ unreleased
 - Update braintree-web to version 3.18.0
 - Update paypal checkout.js to version 4.0.82
 - Allow card overrides with Hosted Fields
+- Use npm scoped version of browser-detection
 
 1.2.0
 ------